Add files via upload
You can encrypt and after that decrypt any file you want. The credential for logging in are "admin" and the password is created at every start of the program in the text file "password_file.txt".
In timer.py you extract the hour, minute, and the date of the moment you start the program.
In password_generator.py it will create a randomize password which contains letters, digits and symbols.
Every password which are created are containing a number of letters equals to the number of the minutes, a number of digits equals to the number of hour and a number of symbols equals to the number of the date.
In Cipher.py is implemented the algorithm of the encryption and decryption process.
Let you to discover more, enjoy it!